Move all ES communication to TextRepoElasticClient
