+++ textrepo-app/src/main/java/nl/knaw/huc/service/health/ElasticsearchHealthCheck.java
@@ -2,48 +2,41 @@ package nl.knaw.huc.service.health;
 
 import com.codahale.metrics.health.HealthCheck;
 import nl.knaw.huc.service.index.TextRepoElasticClient;
-import org.elasticsearch.action.admin.cluster.health.ClusterHealthRequest;
-import org.elasticsearch.action.admin.cluster.health.ClusterHealthResponse;
+import org.elasticsearch.cluster.health.ClusterHealthStatus;
 
 import java.io.IOException;
 
+import static com.codahale.metrics.health.HealthCheck.Result.healthy;
+import static com.codahale.metrics.health.HealthCheck.Result.unhealthy;
 import static java.lang.String.format;
 import static java.util.Arrays.asList;
-import static org.elasticsearch.client.RequestOptions.DEFAULT;
 import static org.elasticsearch.cluster.health.ClusterHealthStatus.GREEN;
 import static org.elasticsearch.cluster.health.ClusterHealthStatus.YELLOW;
 
 public class ElasticsearchHealthCheck extends HealthCheck {
 
-  private final String indexName;
   private final TextRepoElasticClient client;
 
-  public ElasticsearchHealthCheck(String indexName, TextRepoElasticClient client) {
-    this.indexName = indexName;
+  public ElasticsearchHealthCheck(TextRepoElasticClient client) {
     this.client = client;
   }
 
   @Override
   protected Result check() {
-    var request = new ClusterHealthRequest(indexName);
 
-    ClusterHealthResponse response;
+    ClusterHealthStatus status;
     try {
-      response = client
-          .getClient()
-          .cluster()
-          .health(request, DEFAULT);
+      status = client.getHealthStatus();
     } catch (IOException ex) {
-      return HealthCheck.Result.unhealthy(format(
+      return unhealthy(format(
           "Health status: unknown; reason: %s: %s",
           ex.getClass().getName(), ex.getMessage()
       ));
     }
-
     var allowed = asList(GREEN, YELLOW);
-    if (allowed.contains(response.getStatus())) {
-      return HealthCheck.Result.healthy("Health status: " + response.getStatus());
+    if (allowed.contains(status)) {
+      return healthy("Health status: " + status);
     }
-    return HealthCheck.Result.unhealthy("Health status: " + response.getStatus() + "; reason: " + response.status());
+    return unhealthy("Health status: " + status + "; reason: " + status);
   }
 }
