+++ textrepo-app/src/main/java/nl/knaw/huc/service/index/JdbiIndexService.java
@@ -77,61 +77,16 @@ public class JdbiIndexService implements IndexService {
 
   @Override
   public void delete(UUID fileId) {
-    indexClients.forEach(indexClient -> deleteFromIndex(fileId, indexClient));
+    indexClients.forEach(indexClient -> indexClient.delete(fileId));
   }
 
   @Override
   public List<UUID> getAllIds() {
     var result = new HashSet<UUID>();
-    indexClients.forEach(indexClient -> result.addAll(getAllIdsFromIndex(indexClient)));
+    indexClients.forEach(indexClient -> result.addAll(indexClient.getAllIds()));
     return result.stream().toList();
   }
 
-  /**
-   * Retrieve all doc IDs from index using ES scroll api
-   */
-  private List<UUID> getAllIdsFromIndex(TextRepoElasticClient indexClient) {
-    var indexName = indexClient.getConfig().index;
-    try {
-      final var scroll = new Scroll(TimeValue.timeValueMinutes(1L));
-      var searchRequest = new SearchRequest(indexName);
-      searchRequest.scroll(scroll);
-      var searchSourceBuilder = new SearchSourceBuilder();
-      searchSourceBuilder.query(matchAllQuery());
-      searchRequest.source(searchSourceBuilder);
-
-      var client = indexClient.getClient();
-      var searchResponse = client.search(searchRequest, DEFAULT);
-      var scrollId = searchResponse.getScrollId();
-      var result = getIds(searchResponse);
-      var hasHits = result.size() > 0;
-
-      while (hasHits) {
-        var scrollRequest = new SearchScrollRequest(scrollId);
-        scrollRequest.scroll(scroll);
-        searchResponse = client.scroll(scrollRequest, DEFAULT);
-        scrollId = searchResponse.getScrollId();
-        var newHits = getIds(searchResponse);
-        hasHits = newHits.size() > 0;
-        result.addAll(newHits);
-      }
-
-      log.debug("Found {} files in index {}", result.size(), indexName);
-      var clearScrollRequest = new ClearScrollRequest();
-      clearScrollRequest.addScrollId(scrollId);
-      client.clearScroll(clearScrollRequest, DEFAULT);
-      return result;
-    } catch (IOException ex) {
-      throw new WebApplicationException(format("Could not retrieve IDs from index %s", indexName), ex);
-    }
-  }
-
-  private List<UUID> getIds(SearchResponse searchResponse) {
-    return Arrays
-        .stream(searchResponse.getHits().getHits())
-        .map(hit -> UUID.fromString(hit.getId()))
-        .collect(toList());
-  }
 
   private String getLatestVersionContents(TextRepoFile file) {
     var latestVersion = jdbi
@@ -150,28 +105,5 @@ public class JdbiIndexService implements IndexService {
     return latestContents;
   }
 
-  private void deleteFromIndex(@Nonnull UUID fileId, TextRepoElasticClient indexClient) {
-    var indexName = indexClient.getConfig().index;
-    log.info(format("Deleting file %s from index %s", fileId, indexName));
-    DeleteResponse response;
-    var deleteRequest = new DeleteRequest();
-    deleteRequest.index(indexName);
-    deleteRequest.id(fileId.toString());
-    try {
-      response = indexClient.getClient().delete(deleteRequest, DEFAULT);
-    } catch (Exception ex) {
-      throw new WebApplicationException(format("Could not delete file %s in index %s", fileId, indexName), ex);
-    }
-    var status = response.status().getStatus();
-    final String msg;
-    if (status == 200) {
-      msg = format("Successfully deleted file %s from index %s", fileId, indexName);
-    } else if (status == 404) {
-      msg = format("File %s not found in index %s", fileId, indexName);
-    } else {
-      throw new WebApplicationException(format("Could not delete file %s from index %s", fileId, indexName));
-    }
-    log.info(msg);
-  }
 
 }
